+++ src/constant/buttonContent.ts
@@ -1,0 +1,6 @@
+export const ButtonActions = {
+  ACTIVATE: 'addition-option-activeUser',
+  EDIT: 'addition-option-edit',
+  VIEW_PROFILE: 'addition-option-viewProfile',
+  DELETE: 'addition-option-deleteUser',
+};
