+++ src/constant/tableHeader.ts
@@ -1,0 +1,20 @@
+import { HeaderItem } from "@/types/Content";
+
+export const HEADER_ITEMS: HeaderItem[] = [
+  {
+    className: 'name',
+    content: 'NAME',
+  },
+  {
+    className: 'status',
+    content: 'USER STATUS',
+  },
+  {
+    className: 'paymentStatus',
+    content: 'PAYMENT STATUS',
+  },
+  {
+    className: 'amount',
+    content: 'AMOUNT',
+  },
+];
