+++ src/types/Content.ts
@@ -1,0 +1,9 @@
+export interface HeaderItem {
+  className: string;
+  content: string;
+}
+
+
+export interface ActivityHeaderConfig {
+  readonly titles: string[];
+}
