--- src/views/components/index.ts
@@ -1,1 +1,0 @@
-export * from './Header/Header';
