+++ src/views/pages/UserView.ts
@@ -1,12 +1,11 @@
-import { UserResponse } from '../../controllers/users';
 import { TableData } from '../components/Table/TableData/TableData';
 import { Pagination } from '../components/Pagination/Pagination';
 import { UserController } from '@/controllers/users';
 import { FilterManager } from '@/services/FilterManager';
 import { SearchManager } from '@/services/SearchManager';
 import { showNotification } from '@/helpers/showNotification';
-
-type FetchType = 'all' | 'paid' | 'unpaid' | 'overdue';
+import { UserResponse } from '@/types/UserResponse';
+import { FetchType } from '@/types/Filter';
 
 export abstract class UserView {
   protected container: HTMLDivElement;
@@ -70,7 +69,7 @@ export abstract class UserView {
       this.setupSubscriptions();
     } catch (error) {
       console.error('Error fetching data:', error);
-      showNotification('Wait and reload');
+      showNotification('loadError');
     }
   }
 
@@ -107,7 +106,7 @@ export abstract class UserView {
       this.initTable(processedData);
     } catch (error) {
       console.error('Error processing data:', error);
-      showNotification('Wait and reload');
+      showNotification('loadError');
     }
   }
